+++ src/Test/Tensor.hs
@@ -75,7 +75,6 @@ import Prelude hiding (zipWith, replicate)
 
 import Control.Monad.Trans.State (StateT(..), evalStateT)
 import Data.Bifunctor
-import Data.Foldable (foldl')
 import Data.Foldable qualified as Foldable
 import Data.List qualified as L
 import Data.Maybe (catMaybes)
@@ -97,6 +96,15 @@ import Test.QuickCheck qualified as QC
   Definition
 -------------------------------------------------------------------------------}
 
+-- | N-dimensional tensor
+--
+-- Invariants:
+--
+-- * The dimension must be strictly positive (zero is not allowed)
+-- * Tensors must be rectangular
+--
+-- (These invariants could in principle be enforced by using more precise types,
+-- but at the cost of much more complex code.)
 data Tensor n a where
   Scalar :: a -> Tensor Z a
   Tensor :: [Tensor n a] -> Tensor (S n) a
@@ -232,7 +240,7 @@ convolveWithStride stride kernel input =
     aux <$> applyStride stride (subs (size kernel) input)
   where
     aux :: Tensor n a -> a
-    aux = foldl' (+) 0 . zipWith (*) kernel
+    aux = Foldable.foldl' (+) 0 . zipWith (*) kernel
 
 {-------------------------------------------------------------------------------
   Padding