+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for tmp
 
-## 0.1.0.0 -- YYYY-mm-dd
+## 0.1.1 -- 2025-08-27
 
-* First version. Released on an unsuspecting world.
+* Add missing instances for `TestValue`
+* Document `Tensor` invariants (#11)
+
+## 0.1.0 -- 2025-02-15
+
+* First release